app/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
pytest==7.3.1
fastapi==0.95.1 
uvicorn==0.22.0
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.24.0
pathlib==1.0.1